Test TreehouseDispatchers
These initial tests are pretty basic.

Also move TreehouseDispatchers and TreehousePlatform implementations
to their own files.
