+++ redwood-treehouse-host/src/androidMain/kotlin/app/cash/redwood/treehouse/treehouseAppFactoryAndroid.kt
@@ -16,19 +16,11 @@
 package app.cash.redwood.treehouse
 
 import android.content.Context
-import android.os.Looper
-import android.util.Log
 import app.cash.zipline.loader.ManifestVerifier
-import app.cash.zipline.loader.ZiplineCache
 import app.cash.zipline.loader.asZiplineHttpClient
-import java.util.concurrent.Executors
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.asCoroutineDispatcher
 import okhttp3.OkHttpClient
 import okio.FileSystem
 import okio.Path
-import okio.Path.Companion.toOkioPath
 
 public fun TreehouseAppFactory(
   context: Context,
@@ -53,47 +45,3 @@ public fun TreehouseAppFactory(
   cacheMaxSizeInBytes = cacheMaxSizeInBytes,
   concurrentDownloads = concurrentDownloads,
 )
-
-internal class AndroidTreehousePlatform(
-  private val context: Context,
-) : TreehousePlatform {
-  override fun logInfo(message: String, throwable: Throwable?) {
-    Log.i("Zipline", message, throwable)
-  }
-
-  override fun logWarning(message: String, throwable: Throwable?) {
-    Log.w("Zipline", message, throwable)
-  }
-
-  override fun newCache(name: String, maxSizeInBytes: Long) = ZiplineCache(
-    context = context,
-    fileSystem = FileSystem.SYSTEM,
-    directory = context.cacheDir.toOkioPath() / name,
-    maxSizeInBytes = maxSizeInBytes,
-  )
-}
-
-/**
- * Implements [TreehouseDispatchers] suitable for production Android use. This creates a background
- * thread for all Zipline work.
- */
-internal class AndroidTreehouseDispatchers : TreehouseDispatchers {
-  private lateinit var ziplineThread: Thread
-
-  /** The single thread that runs all JavaScript. We only have one QuickJS instance at a time. */
-  private val executorService = Executors.newSingleThreadExecutor { runnable ->
-    Thread(runnable, "Treehouse")
-      .also { ziplineThread = it }
-  }
-
-  override val ui: CoroutineDispatcher get() = Dispatchers.Main
-  override val zipline: CoroutineDispatcher = executorService.asCoroutineDispatcher()
-
-  override fun checkUi() {
-    check(Looper.myLooper() == Looper.getMainLooper())
-  }
-
-  override fun checkZipline() {
-    check(Thread.currentThread() == ziplineThread)
-  }
-}
