+++ redwood-treehouse-host/src/commonTest/kotlin/app/cash/redwood/treehouse/AbstractFrameClockTest.kt
@@ -36,6 +36,7 @@ abstract class AbstractFrameClockTest {
       override val zipline get() = ui
       override fun checkUi() {}
       override fun checkZipline() {}
+      override fun close() {}
     }
     frameClock.start(this, dispatchers)
