+++ redwood-treehouse-host/src/iosMain/kotlin/app/cash/redwood/treehouse/IosTreehousePlatform.kt
@@ -1,0 +1,46 @@
+/*
+ * Copyright (C) 2023 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package app.cash.redwood.treehouse
+
+import app.cash.zipline.loader.ZiplineCache
+import okio.FileSystem
+import okio.Path.Companion.toPath
+import platform.Foundation.NSLog
+import platform.Foundation.NSTemporaryDirectory
+
+internal class IosTreehousePlatform : TreehousePlatform {
+  override fun logInfo(message: String, throwable: Throwable?) {
+    if (throwable != null) {
+      NSLog("Treehouse: $message ${throwable.stackTraceToString()}")
+    } else {
+      NSLog("Treehouse: $message")
+    }
+  }
+
+  override fun logWarning(message: String, throwable: Throwable?) {
+    if (throwable != null) {
+      NSLog("Treehouse: $message ${throwable.stackTraceToString()}")
+    } else {
+      NSLog("Treehouse: $message")
+    }
+  }
+
+  override fun newCache(name: String, maxSizeInBytes: Long) = ZiplineCache(
+    fileSystem = FileSystem.SYSTEM,
+    directory = NSTemporaryDirectory().toPath() / name,
+    maxSizeInBytes = maxSizeInBytes,
+  )
+}
