+++ redwood-treehouse-host/src/iosTest/kotlin/app/cash/redwood/treehouse/IosTreehouseDispatchersTest.kt
@@ -1,0 +1,24 @@
+/*
+ * Copyright (C) 2023 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package app.cash.redwood.treehouse
+
+class IosTreehouseDispatchersTest : AbstractTreehouseDispatchersTest() {
+  override val treehouseDispatchers: TreehouseDispatchers = IosTreehouseDispatchers()
+
+  /** We haven't set done the work to dispatch to the UI thread on iOS tests. */
+  override val ignoreTestsThatExecuteOnUiThread: Boolean
+    get() = true
+}
